requirements.txt
asgiref==3.5.0
Django==4.0.1
django-phonenumber-field==6.0.0
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
phonenumbers==8.12.42
psycopg-binary==3.0.8
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
